pyproject.toml
[project]
name = "dot"
version = "0.1.1"
description = "Data Orchestration Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "pygit2>=1.14.1",
]

[project.scripts]
dot = "dot.cli:app"
